+++ index.html
@@ -7,11 +7,6 @@
       type="image/svg+xml"
       href="./src/shared/assets/images/images/icon.svg"
     />
-    <meta
-      http-equiv="Content-Security-Policy"
-      content="upgrade-insecure-requests"
-    />
-
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>React</title>
   </head>
