+++ src/app/providers/routerProvider/index.ts
@@ -1,0 +1,1 @@
+export { RouteProvider } from './ui/routeProvider';
