+++ src/app/providers/routerProvider/ui/routeProvider.tsx
@@ -1,0 +1,19 @@
+import { ReactNode } from 'react';
+import { Route, Routes } from 'react-router-dom';
+import { rootRouter, Router } from '../config/rootRouter';
+
+export const RouteProvider = (): ReactNode => {
+  return (
+    <Routes>
+      (
+      {rootRouter.map((route: Router) => (
+        <Route
+          key={route.path}
+          path={route.path}
+          element={route.element}
+        ></Route>
+      ))}
+      )
+    </Routes>
+  );
+};
