+++ src/index.tsx
@@ -1,9 +1,12 @@
 import { createRoot } from 'react-dom/client';
 import React from 'react';
 import { App } from 'app/App';
+import { BrowserRouter } from 'react-router-dom';
 
 createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
-    <App />
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
   </React.StrictMode>
 );
