+++ src/pages/Main/Main.module.scss
@@ -1,4 +1,4 @@
-@import './common.scss';
+@import '../../app/styles/common.scss';
 
 .species {
   background: url('../../shared/assets/images/background/species.jpg');
