+++ src/pages/NotFound/NotFound.tsx
@@ -1,0 +1,5 @@
+import { FC } from 'react';
+
+export const NotFound: FC = () => {
+  return <></>;
+};
