+++ src/shared/types/routerTypes.ts
@@ -1,0 +1,4 @@
+export enum Paths {
+  MAIN = '/',
+  NOT_FOUND = '*',
+}
